PostProcessing added + joystick
Now we have postprocessing, it is working like this: There is a camera in the player "PlayerCamera", it renders to a custom render texture, that i modify with some hlsl, after that i display that texture to a rawImage in the canvas, called "PostProcessing", you can find this as a prefab.
How to make it work? Just put the "player" prefab into the scene, add the "PostProcessing" prefab as well (this is canvas). Ensure that the Main Camera stays free from parents. (If you leave the Main camera as she was when we create a new Scene, it is going to work fine)